temperature is now float. shaking code linted.

+++ db/shaking.go
@@ -11,11 +11,11 @@ import (
 type Shaking struct {
 	ID          uuid.UUID     `json:"id" db:"id"`
 	WithTemp    bool          `json:"with_temp" db:"with_temp"`
-	Temperature int64           `json:"temperature" db:"temperature"`
+	Temperature float64       `json:"temperature" db:"temperature"`
 	FollowTemp  bool          `json:"follow_temp" db:"follow_temp"`
 	ProcessID   uuid.UUID     `json:"process_id" db:"process_id"`
-	RPM1        int64           `json:"rpm_1" db:"rpm_1"`
-	RPM2        int64           `json:"rpm_2" db:"rpm_2"`
+	RPM1        int64         `json:"rpm_1" db:"rpm_1"`
+	RPM2        int64         `json:"rpm_2" db:"rpm_2"`
 	Time1       time.Duration `json:"time_1" db:"time_1"`
 	Time2       time.Duration `json:"time_2" db:"time_2"`
 	CreatedAt   time.Time     `json:"created_at" db:"created_at"`
@@ -23,7 +23,7 @@ type Shaking struct {
 }
 
 const (
-	getShakingQuery = `SELECT * FROM shaking where process_id = $1`
+	getShakingQuery    = `SELECT * FROM shaking where process_id = $1`
 	createShakingQuery = `INSERT INTO shaking (
 		with_temp,
 		temperature,
@@ -37,7 +37,7 @@ const (
 )
 
 func (s *pgStore) ShowShaking(ctx context.Context, shakingID uuid.UUID) (shaking Shaking, err error) {
-	
+
 	err = s.db.Get(&shaking, getShakingQuery, shakingID)
 	if err != nil {
 		logger.WithField("err", err.Error()).Error("Error getting shaking data")
@@ -49,7 +49,7 @@ func (s *pgStore) ShowShaking(ctx context.Context, shakingID uuid.UUID) (shaking
 
 func (s *pgStore) CreateShaking(ctx context.Context, sh Shaking) (createdShaking Shaking, err error) {
 	var lastInsertID uuid.UUID
-	
+
 	err = s.db.QueryRow(
 		createShakingQuery,
 		sh.WithTemp,
